bmq.dev.configurator: split and rework dependencies
