+++ src/integration-tests/test_graceful_shutdown.py
@@ -16,7 +16,7 @@ from blazingmq.dev.it.fixtures import (  # pylint: disable=unused-import
 )
 from blazingmq.dev.it.process.client import Client
 from blazingmq.dev.it.util import wait_until
-from blazingmq.dev.configurator import Configurator
+from blazingmq.dev.configurator.configurator import Configurator
 
 OTHER_DOMAIN = f"{tc.DOMAIN_PRIORITY}.other"
