+++ src/python/blazingmq/dev/it/cluster.py
@@ -7,12 +7,13 @@ import itertools
 import logging
 import shutil
 import signal
-from typing import Any, Dict, List, Optional, Union
+from typing import Dict, List, Optional, Union
 from pathlib import Path
 
+from blazingmq.dev.configurator.localsite import LocalSite
 import blazingmq.dev.it.process.proc
 import blazingmq.dev.it.testconstants as tc
-import blazingmq.dev.configurator as cfg
+import blazingmq.dev.configurator.configurator as cfg
 from blazingmq.dev.it.process.broker import Broker
 from blazingmq.dev.it.process.client import Client
 from blazingmq.dev.it.util import ListContextManager, Queue, internal_use
@@ -594,7 +595,9 @@ class Cluster(contextlib.AbstractContextManager):
 
     def deploy_domains(self):
         for broker in self.configurator.brokers.values():
-            broker.deploy_domains(cfg.LocalSite(self.work_dir / broker.name))
+            self.configurator.deploy_domains(
+                broker, LocalSite(self.work_dir / broker.name)
+            )
 
     def reconfigure_domain(
         self,
